+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/open-telemetry/opentelemetry-python/compare/v1.11.1-0.30b1...HEAD)
 
+- Update to semantic conventions v1.11.0
+  ([#2669](https://github.com/open-telemetry/opentelemetry-python/pull/2669))
 - Update opentelemetry-proto to v0.17.0
   ([#2668](https://github.com/open-telemetry/opentelemetry-python/pull/2668))
 - Add CallbackOptions to observable instrument callback params